Add foreach method for linkedlist
Iterator object doesn't seem possible for this design due to data being
inside a refcell. Borrowing it creates a new object which would be
dropped before the reference is returned

It might be possible with another struct holding a Ref that would then
Deref it to a reference. The foreach method should suffice for now